+++ src/main/java/site/siredvin/progressiveperipherals/integrations/computercraft/turtles/TurtleCuttingAxe.java
@@ -20,6 +20,7 @@ import org.jetbrains.annotations.NotNull;
 import site.siredvin.progressiveperipherals.ProgressivePeripherals;
 import site.siredvin.progressiveperipherals.common.configuration.ProgressivePeripheralsConfig;
 import site.siredvin.progressiveperipherals.common.setup.Items;
+import site.siredvin.progressiveperipherals.integrations.computercraft.turtles.base.TurtleDigOperationType;
 import site.siredvin.progressiveperipherals.integrations.computercraft.turtles.base.TurtleDigTool;
 import site.siredvin.progressiveperipherals.utils.TranslationUtil;
 
@@ -54,6 +55,11 @@ public class TurtleCuttingAxe extends TurtleDigTool {
     }
 
     @Override
+    public TurtleDigOperationType getOperationType() {
+        return TurtleDigOperationType.CUTTING_AXE;
+    }
+
+    @Override
     protected TurtleCommandResult dig(@NotNull ITurtleAccess turtle, @NotNull TurtleSide side, @NotNull Direction direction) {
         World world = turtle.getWorld();
         BlockPos turtlePosition = turtle.getPosition();
@@ -74,6 +80,8 @@ public class TurtleCuttingAxe extends TurtleDigTool {
             ProgressivePeripherals.LOGGER.info("Tree cutting stopped because of max size");
         TurtlePlayer turtlePlayer = TurtlePlayer.getWithPosition(turtle, turtlePosition, direction);
         turtlePlayer.loadInventory(mimicTool());
+        if (!consumeFuel(turtle))
+            return TurtleCommandResult.failure("Not enough fuel");
         for (BlockPos pos: treeBlocks) {
             digOneBlock(turtle, side, world, pos, turtlePlayer, turtleTile);
         }
@@ -114,16 +122,17 @@ public class TurtleCuttingAxe extends TurtleDigTool {
         }
     }
 
-    public static EnchantedTurtleCuttingAxe enchant(String prefix, Enchantment enchantment, int enchantmentLevel) {
-        return new EnchantedTurtleCuttingAxe(prefix, enchantment, enchantmentLevel);
+    public static EnchantedTurtleCuttingAxe enchant(String prefix, Enchantment enchantment, int enchantmentLevel, TurtleDigOperationType newOperationType) {
+        return new EnchantedTurtleCuttingAxe(prefix, enchantment, enchantmentLevel, newOperationType);
     }
 
     public static class EnchantedTurtleCuttingAxe extends TurtleCuttingAxe {
 
         private final Enchantment enchantment;
         private final int enchantmentLevel;
+        private final TurtleDigOperationType operationType;
 
-        public EnchantedTurtleCuttingAxe(String prefix, Enchantment enchantment, int enchantmentLevel) {
+        public EnchantedTurtleCuttingAxe(String prefix, Enchantment enchantment, int enchantmentLevel, TurtleDigOperationType newOperationType) {
             super(new ResourceLocation(ProgressivePeripherals.MOD_ID, prefix + CORE_NAME), TranslationUtil.turtle(prefix + CORE_NAME), () -> {
                 ItemStack craftingItem = new ItemStack(Items.CUTTING_AXE.get());
                 craftingItem.enchant(enchantment, enchantmentLevel);
@@ -131,6 +140,12 @@ public class TurtleCuttingAxe extends TurtleDigTool {
             });
             this.enchantment = enchantment;
             this.enchantmentLevel = enchantmentLevel;
+            this.operationType = newOperationType;
+        }
+
+        @Override
+        public TurtleDigOperationType getOperationType() {
+            return operationType;
         }
 
         @Override